Switch to using pull_request_summarizer gem
