+++ Gemfile
@@ -6,9 +6,9 @@ gem 'pry'
 gem 'octokit'
 gem 'rack-github_webhooks'
 gem 'rake'
-gem 'everypolitician-popolo', git: 'https://github.com/everypolitician/everypolitician-popolo', branch: 'master'
 gem 'rubocop', '~> 0.42.0'
-gem 'require_all'
+gem 'pull_request_summarizer'
+gem 'dotenv'
 
 group :test do
   gem 'minitest'
