--- comment_template.md.erb
@@ -1,126 +1,0 @@
-<% if popolo_files.empty? %>
-No Popolo files were changed in this pull request.
-<% else %>
-<% popolo_files.each do |file| %>
-Summary of changes in `<%= file.path %>`:
-
-## People
-
-### Added
-
-<% if file.people_added.any? %>
-<% file.people_added.each do |person| %>
-- `<%= person.id %>` - <%= person.name %>
-<% end %>
-<% else %>
-No people added
-<% end %>
-
-### Removed
-
-<% if file.people_removed.any? %>
-<% file.people_removed.each do |person| %>
-- `<%= person.id %>` - <%= person.name %>
-<% end %>
-<% else %>
-No people removed
-<% end %>
-
-### Name Changes
-
-<% if file.people_name_changes.any? %>
-<% file.people_name_changes.each do |r| %>
-- `<%= r[:id] %>`: <%= r[:was] %> → <%= r[:now] %>
-<% end %>
-<% else %>
-No name changes
-<% end %>
-
-### Additional Name Changes
-
-<% if file.people_additional_name_changes.any? %>
-<% file.people_additional_name_changes.each do |r| %>
-- `<%= r[:id] %>` (<%= r[:name] %>): <% if r[:removed].any? %>Removed: <%= r[:removed].join(" ﹠ ") %>. <% end %><% if r[:added].any? %>Added: <%= r[:added].join(" ﹠ ") %>.<% end %>
-<% end %>
-<% else %>
-No name changes
-<% end %>
-
-### Wikidata Changes
-
-<% if file.wikidata_links_changed.any? %>
-<% file.wikidata_links_changed.each do |r| %>
-- `<%= r[:id] %>`: <%= r[:was] %> → <%= r[:now] %>
-<% end %>
-<% else %>
-No changes
-<% end %>
-
-## Organizations
-
-### Added
-
-<% if file.organizations_added.any? %>
-<% file.organizations_added.each do |org| %>
-- `<%= org.id %>` - <%= org.name %>
-<% end %>
-<% else %>
-No organizations added
-<% end %>
-
-### Removed
-
-<% if file.organizations_removed.any? %>
-<% file.organizations_removed.each do |org| %>
-- `<%= org.id %>` - <%= org.name %>
-<% end %>
-<% else %>
-No organizations removed
-<% end %>
-
-## Terms
-
-### Added
-
-<% if file.terms_added.any? %>
-<% file.terms_added.each do |term| %>
-- `<%= term.id %>` - <%= term.name %>
-<% end %>
-<% else %>
-No terms added
-<% end %>
-
-### Removed
-
-<% if file.terms_removed.any? %>
-<% file.terms_removed.each do |term| %>
-- `<%= term.id %>` - <%= term.name %>
-<% end %>
-<% else %>
-No terms removed
-<% end %>
-
-## Elections
-
-### Added
-
-<% if file.elections_added.any? %>
-<% file.elections_added.each do |election| %>
-- `<%= election.id %>` - <%= election.name %>
-<% end %>
-<% else %>
-No elections added
-<% end %>
-
-### Removed
-
-<% if file.elections_removed.any? %>
-<% file.elections_removed.each do |election| %>
-- `<%= election.id %>` - <%= election.name %>
-<% end %>
-<% else %>
-No elections removed
-<% end %>
-
-<% end %>
-<% end %>
