--- lib/report/people.rb
@@ -1,11 +1,0 @@
-class Report
-  class People < Base
-    def added
-      after.persons - before.persons
-    end
-
-    def removed
-      before.persons - after.persons
-    end
-  end
-end
