--- test/fixtures/after.json
@@ -1,50 +1,0 @@
-{
-  "persons": [
-    {
-      "id": "ef41232b-daa4-40fd-a39f-1ff900e3b1c2"
-    },
-    {
-      "id": "c74ad586-c6b1-4e9b-92cd-23babc136856"
-    },
-    {
-      "id": "83af0ada-521a-4043-8b14-49dff73a9389"
-    }
-  ],
-  "organizations": [
-    {
-      "id": "4d814738-39d3-4a09-835d-32a2ed47c3c3"
-    },
-    {
-      "id": "47c71c45-e37e-4936-ba3d-e4abee929da0"
-    },
-    {
-      "id": "47e60c56-a663-4cad-b8d8-b63309fdd7c4"
-    }
-  ],
-  "events": [
-    {
-      "classification": "legislative period",
-      "id": "term/53",
-      "name": "53rd Parliament of the United Kingdom"
-    },
-    {
-      "classification": "legislative period",
-      "id": "term/54",
-      "name": "54th Parliament of the United Kingdom"
-    },
-    {
-      "classification": "general election",
-      "end_date": "1996-11-23",
-      "id": "Q4668016",
-      "name": "Abkhazian parliamentary election, 1996",
-      "start_date": "1996-11-23"
-    },
-    {
-      "classification": "general election",
-      "end_date": "2002-03-02",
-      "id": "Q16960120",
-      "name": "Abkhazian parliamentary election, 2002",
-      "start_date": "2002-03-02"
-    }
-  ]
-}
